app/services/web/requirements.txt
werkzeug==2.2.2
markupsafe
Flask
gunicorn
gevent
itsdangerous
psycopg2-binary
Flask-SQLAlchemy==3.0.3
redis
celery
click
Pillow
requests
pywebpush
python-dateutil
pandas

Flask-WTF
WTForms-Alchemy
WTForms-Components

flask-debugtoolbar
Flask-Mail
Flask-Login
flask-bcrypt
flask-migrate
flask-cors
flask-script
Flask-Limiter
WTForms-Alchemy

google-cloud-storage
google-cloud-bigquery

